done render lvl 2

+++ Game/core/src/org/academiadecodigo/hashtronauts/levels/gameObjects/ExitPoint.java
@@ -2,7 +2,7 @@ package org.academiadecodigo.hashtronauts.levels.gameObjects;
 
 public class ExitPoint extends GameObject {
 
-    public ExitPoint(String texture, int posX, int posY, int width, int height) {
+    public ExitPoint(String texture, float posX, float posY, int width, int height) {
         super(texture, posX, posY, width, height);
     }
